extracted the label logic to a module
