+++ data/authors.json
@@ -1,1 +1,1 @@
-[{"id":865,"first_name":"James","last_name":"Clear"}]
+[{"id":566,"first_name":"James","last_name":"Clear"}]
