+++ data/books.json
@@ -1,1 +1,1 @@
-{"data":[{"id":712,"genre":314,"author":865,"label":106,"publish_date":"2006-01-01","publisher":"Medium","cover_state":"bad"},{"id":70,"genre":797,"author":865,"label":821,"publish_date":"3021-12-12","publisher":"medium","cover_state":"good"}]}
+{"data":[{"id":132,"genre":454,"author":566,"label":633,"publish_date":"2009-12-12","publisher":"medium","cover_state":"bad"}]}
