+++ data/genres.json
@@ -1,1 +1,1 @@
-[{"id":314,"name":"Comedy","items":[]},{"id":797,"name":"Thriller","items":[]}]
+[{"id":454,"name":"Comedy","items":[]}]
