+++ data/label.json
@@ -1,1 +1,1 @@
-{"data":[{"id":106,"title":"Gift","color":"unknown"},{"id":821,"title":"New","color":"unknown"}]}
+{"data":[{"id":633,"title":"Gift","color":"unknown"}]}
